add optimized logo
this is a scaled-down, optimized version, intended very specifically to ONLY be used for the small logo in the upper-left-hand corner. if you need an unoptimized PNG, get in touch with me (Gabe). if you need access to the original Illustrator files and the logos Google Drive folder, get in touch with marketing.